feat(MyLocation): disable follow location on place or hiking route details load

+++ app/src/main/java/hu/mostoha/mobile/android/turistautak/extensions/MapViewExtensions.kt
@@ -139,8 +139,8 @@ fun MapView.addFastOverlay(
     return fastOverlay
 }
 
-fun MapView.centerAndZoom(geoPoint: GeoPoint, zoomLevel: Int) {
-    controller.setZoom(zoomLevel.toDouble())
+fun MapView.centerAndZoom(geoPoint: GeoPoint, zoomLevel: Double) {
+    controller.setZoom(zoomLevel)
     controller.setCenter(geoPoint)
 }
